requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
moviepy==1.0.3
openai==1.3.7
whisper==1.0
boto3==1.34.34
google-api-python-client==2.105.0
google-auth-oauthlib==1.1.0
requests==2.31.0
apscheduler==3.10.4
redis==5.0.1
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
python-dotenv==1.0.0
prometheus-fastapi-instrumentator==6.1.0
slowapi==0.1.9
